chore: move constants module to data directory as this file does not contain any code
