+++ dev_scripts/create_end_to_end_test_json.py
@@ -19,7 +19,7 @@ from typing import Tuple, Dict, Union, List
 import numpy as np
 import pandas as pd
 
-import nasem_dairy.model.constants as constants
+import nasem_dairy.data.constants as constants
 
 
 ####################
